fix(mobile): move home contents to index, remove unused code
