+++ mobile/searchitect/app/index.tsx
@@ -1,46 +1,95 @@
-import { Stack, useNavigation } from 'expo-router';
-import { Text, View } from 'react-native';
-import { useEffect } from 'react';
-import QuizDetails from './quizDetails';
-import QuizFeed from './quizFeed';
-
-import QuizQuestion from './quizQuestion';
-import QuizCreationSettings from './quizCreationSettings';
-import QuizCreationInfo from './quizCreationInfo';
-import QuizCreationQuestionList from './quizCreationQuestionList';
-
-// import Authentication from './auth';
+import React, {useState} from 'react';
+import {Image, Pressable, StyleSheet, Text, View} from 'react-native';
+import { router } from "expo-router";
+import Navbar from "./navbar";
 
 export default function Home() {
-    console.log("index.tsx run")
-
-  const navigation = useNavigation();
-
-  useEffect(() => {
-    navigation.setOptions({ headerShown: false });
-  }, [navigation]);
-
   return (
-    <>
-      <QuizQuestion question={'QUESTIONNNN'} options={["aaa","bbb","ccc","ddd"]} currentQuestion={0} totalQuestions={0}/>
-</>
+    <View style={styles.container}>
+      <Navbar/>
+      <View style={styles.page}>
+      <View style={styles.profilePictureContainer}>
+      <Image
+            source={require('@/assets/images/profile-icon.png')}
+            style={styles.profilePicture}
+        />
+      </View>
+      <View style={styles.messageContainer}>
+        <Text style={styles.bigText}>Welcome!</Text>
+        <Text style={styles.text}>You are not logged in. Register or Log In to access all features.</Text>
+      </View>
+      <Pressable style={styles.rectangularButton}>
+        <Text style={styles.buttonText}>Register</Text>
+      </Pressable>
+      <Pressable style={styles.rectangularButton} onPress={() => { router.navigate("/login") }}>
+        <Text style={styles.buttonText}>Log In</Text>
+      </Pressable>
+      </View>
+    </View>
   );
 }
 
 
-
-
-// <Link
-// href={{
-//     pathname: "/details",
-//     // /* 1. Navigate to the details route with query params */
-//     params: { id: 86, other: "anything you want here" },
-// }}>
-// Go to Details
-// </Link>
-
-{/* <Link href="/authentication/register" asChild>
-<Pressable>
- <Text> REGISTER</Text>
-</Pressable>
-</Link> */}
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "flex-start",
+    padding: 20,
+    backgroundColor: '#fff'
+  },
+  page: {
+    flex: 1,
+    justifyContent: "flex-start",
+    alignItems: 'center',
+    padding: 20,
+    backgroundColor: '#fff'
+  },
+  profilePictureContainer:{
+    justifyContent: 'center',
+    margin: 10,
+    height: 200,
+    alignItems: 'center',
+  },
+  messageContainer:{
+    borderColor: '#000',
+    borderWidth: 3,
+    borderRadius: 16,
+    width: "90%",
+    height: 120,
+    padding: 3,
+    backgroundColor: '#b2f2bb',
+    margin: 10,
+  },
+  profilePicture: {
+    width: 192,
+    height: 192,
+    borderWidth: 4,
+    borderColor: '#000',
+    borderRadius: 96,
+  },
+  bigText: {
+    fontSize: 24,
+    textAlign: 'center',
+  },
+  text: {
+    fontSize: 16,
+    margin: 3,
+    textAlign: 'center',
+  },
+  rectangularButton: {
+    backgroundColor: "#9775fa",
+    width: '90%',
+    height: 50,
+    borderRadius: 10,
+    justifyContent: 'center',
+    alignItems: 'center',
+    margin: 5,
+    borderWidth: 3,
+    borderColor: "#222",
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    fontSize: 22,
+  },
+});
